Move pydantic models to infra/data_collection
